Cleanup Shutdown Thread Handling
It seems like from the beginning Done and Wait were switched.
It feels like the main thread should wait for the side thread to close
down, rather than the side thread waiting on the main thread?
Once the main thread ended, the side thread was killed anyway.

And also on each signal we would close `shutdown`, but it could only be
closed once, so there wasn't really a reason for this to be a loop.
So now it just blocks, then closes the thread once.